+++ src/index.tsx
@@ -21,30 +21,36 @@ import Report from "./page/Report";
 import UserPage from "./page/UserPage";
 import Search from "./page/search";
 import AdminRoute from "./until/AdminRoute";
+import PatchDocument from "./page/patchDocument";
+import TokenRefresher from "./until/check";
 
 const rootElement = document.getElementById('root') as HTMLElement;
 const root = ReactDOM.createRoot(rootElement);
 
+
 root.render(
     <RecoilRoot>
         <BrowserRouter>
-            <Routes>
-                <Route path={'/'} element={<Main />}></Route>
-                <Route path={'/login'} element={<Login />}></Route>
-                <Route path={'/signup'} element={<Signup />}></Route>
-                <Route path='/more' element={<MoreContents />}></Route>
-                <Route path={'/write'} element={<Write />}></Route>
-                <Route path={'/post/:id'} element={<Post />}></Route>
-                <Route path={'/report'} element={<Report />}></Route>
-                <Route path={'/user/:username'} element={<UserPage/>}></Route>
-                <Route path={'/search/:title'} element={<Search />}></Route>
-                <Route element={<AdminRoute />}> {/* admin route*/}
-                    <Route path={'/report/manage'} element={<ReportManage />}></Route>
-                    <Route path={'/report/manage/:id'} element={<DetailedReport />}></Route>
-                    <Route path={'/ban'} element={<BanManage />}></Route>
-                    <Route path={'/ban/user'} element={<UserBan />}></Route>
-                </Route>
-            </Routes>
+            <TokenRefresher>
+                <Routes>
+                    <Route path={'/'} element={<Main />}></Route>
+                    <Route path={'/login'} element={<Login />}></Route>
+                    <Route path={'/signup'} element={<Signup />}></Route>
+                    <Route path='/more' element={<MoreContents />}></Route>
+                    <Route path={'/write'} element={<Write />}></Route>
+                    <Route path={'/post/:id'} element={<Post />}></Route>
+                    <Route path={'/report'} element={<Report />}></Route>
+                    <Route path={'/user/:username'} element={<UserPage/>}></Route>
+                    <Route path={'/search/:title'} element={<Search />}></Route>
+                    <Route path={'/patch/:id'} element={<PatchDocument />}></Route>
+                    <Route element={<AdminRoute />}> {/* admin route*/}
+                        <Route path={'/report/manage'} element={<ReportManage />}></Route>
+                        <Route path={'/report/manage/:id'} element={<DetailedReport />}></Route>
+                        <Route path={'/ban'} element={<BanManage />}></Route>
+                        <Route path={'/ban/user'} element={<UserBan />}></Route>
+                    </Route>
+                </Routes>
+            </TokenRefresher>
         </BrowserRouter>
     </RecoilRoot>
 );
